pyproject.toml
[project]
name = "expense-tracker-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "pydantic-settings",
    "alembic>=1.18.4",
    "fastapi>=0.131.0",
    "uvicorn[standard]",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
]
